+++ test/cases/angular2-ngIf/angular2-ngIf-component.js
@@ -1,3 +1,0 @@
-@View({
-    templateUrl: `angular2-ngIf-template.html`
-})