+++ test/cases/angular2-ngIf/directive.js
@@ -1,0 +1,3 @@
+@View({
+    templateUrl: `template.html`
+})