+++ test/cases/hard-attributes/directive.js
@@ -1,6 +1,6 @@
 angular.module('test').directive('helloWorld', function () {
     return {
         restrict: 'E',
-        templateUrl: 'hard-attributes-template.html'
+        templateUrl: 'template.html'
     };
 });