+++ test/cases/hello-world/directive.js
@@ -1,6 +1,6 @@
 angular.module('test').directive('helloWorld', function () {
     return {
         restrict: 'E',
-        templateUrl: 'hello-world-template.html'
+        templateUrl: 'template.html'
     };
 });