+++ test/cases/max-size/embedded.js
@@ -1,0 +1,6 @@
+angular.module('test').directive('helloWorld', function () {
+    return {
+        restrict: 'E',
+        templateUrl: 'template.html'
+    };
+});