+++ test/cases/skip-errors/directive.js
@@ -1,0 +1,23 @@
+// several directives in one file
+var testModule = angular.module('test');
+
+testModule.directive('helloWorld1', function () {
+    return {
+        restrict: 'E',
+        templateUrl: 'template1.html'
+    };
+});
+
+testModule.directive('helloWorld2', function () {
+    return {
+        restrict: 'E',
+        templateUrl: 'not-existing-template.html'
+    };
+});
+
+testModule.directive('helloWorld3', function () {
+    return {
+        restrict: 'E',
+        templateUrl: 'template3.html'
+    };
+});