+++ test/cases/skip-file/directive.js
@@ -1,0 +1,6 @@
+// skip me
+function prepareAngularTemplate() {
+    return {
+        templateUrl: 'template.html'
+    };
+}